feat: clean up agents. Break large modules into smaller modules.
